refactor: segregated shop views
